Fix holding down enter with "step"
This also makes the prompt show up after
the step finishes if it finishes quickly instead
of slipping in between the request and
response.

Also, if the stepping takes longer than 3 sec
it outputs how long ago the step was
requested, which can be interesting when
you're stepping one thread that then blocks
on a mutex and then you step the thread
that then releases the mutex, you get both
outputs and can understand what's what.